requirements.txt
asgiref==3.7.2
Django==3.2.20
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-shopping-cart==0.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
